requirements.txt
pyserial==3.4
unittest2==1.1.0
pycrypto==2.6.1

